Rename default_sc sources for consistency
