+++ src/default_sc/BIB_HMAC_SHA2.c
