+++ src/default_sc/DefaultSecContext.c
