+++ src/default_sc/DefaultSecContext_Private.h
