+++ src/mock_bpa/agent.c
@@ -27,7 +27,7 @@
 #include <BPSecLib_Public.h>
 #include <BPSecLib_Private.h>
 #include <backend/UtilDefs_SeqReadWrite.h>
-#include <security_context/DefaultSecContext.h>
+#include <default_sc/DefaultSecContext.h>
 #include <policy_provider/SamplePolicyProvider.h>
 #include <errno.h>
 #include <poll.h>
