+++ test/test_AbsSecBlock.c
@@ -28,9 +28,6 @@
 #include <CryptoInterface.h>
 
 #include <backend/PublicInterfaceImpl.h>
-#include <security_context/DefaultSecContext.h>
-#include <security_context/DefaultSecContext_Private.h>
-#include <security_context/rfc9173.h>
 
 #include "bsl_test_utils.h"
